+++ practice/web-diary/src/app-launcher/app/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@app-launcher/app",
+  "version": "0.0.0",
+  "license": "BSD 3-Clause",
+  "main": "src/index.tsx",
+  "scripts": {
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "react": "^17.0.1",
+    "@ui/app-header": "0.0.0",
+    "@ui/post-add-form": "0.0.0",
+    "@ui/post-list": "0.0.0",
+    "@ui/post-list-item": "0.0.0",
+    "@ui/filter": "0.0.0",
+    "@ui/search": "0.0.0",
+    "@app-launcher/app": "0.0.0",
+    "@style/css": "0.0.0"
+  }
+}
