+++ practice/web-diary/src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './components/app/src'
+import App from '@app-launcher/app'
 
-import './style/index.css'
+import './style/src/index.css'
 
 ReactDOM.render(
   <React.StrictMode>
