+++ practice/web-diary/src/style/src/post-list-item.css
