+++ practice/web-diary/src/style/src/post-list.css
