+++ practice/web-diary/src/ui/app-header/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "@components/app-header",
+  "name": "@ui/app-header",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "scripts": {
     "lint": "eslint"
   },
   "dependencies": {
-    "react": "^17.0.1"
-
+    "react": "^17.0.1",
+    "@style/css": "0.0.0"
   }
 }
