+++ practice/web-diary/src/ui/button/package.json
@@ -2,11 +2,12 @@
   "name": "@ui/button",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "scripts": {
     "lint": "eslint"
   },
   "dependencies": {
-    "react": "^17.0.1"
+    "react": "^17.0.1",
+    "@style/css": "0.0.0"
   }
 }
