+++ practice/web-diary/src/ui/post-add-form/src/PostAddForm.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-import '../../../style/post-add-form.css'
-import { InputNewForm } from '../../../ui/input/src'
-import { ButtonSubmit } from '../../../ui/button/src'
+import '../../../style/src/post-add-form.css'
+import { InputNewForm } from '@ui/input'
+import { ButtonSubmit } from '@ui/button'
 
 const PostAddForm = () => {
   return (
