+++ practice/web-diary/src/ui/post-list-item/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "@components/search",
+  "name": "@ui/post-list-item",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "scripts": {
     "lint": "eslint"
   },
   "dependencies": {
     "react": "^17.0.1",
-    "@ui/input": "0.0.0"
+    "@ui/button": "0.0.0",
+    "@style/css": "0.0.0"
   }
 }
