+++ practice/web-diary/src/ui/post-list-item/src/index.tsx
