+++ practice/web-diary/src/ui/post-list/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@components/post-add-form",
+  "name": "@ui/post-list",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "scripts": {
     "lint": "eslint"
   },
   "dependencies": {
     "react": "^17.0.1",
-    "@ui/button": "0.0.0",
-    "@ui/input": "0.0.0"
+    "@ui/post-list-item": "0.0.0",
+    "@style/css": "0.0.0"
   }
 }
