+++ practice/web-diary/src/ui/post-status-filter/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "@components/filter",
+  "name": "@ui/filter",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "scripts": {
     "lint": "eslint"
   },
   "dependencies": {
     "react": "^17.0.1",
-    "@ui/button": "0.0.0"
+    "@ui/button": "0.0.0",
+    "@style/css": "0.0.0"
   }
 }
