+++ practice/web-diary/src/ui/search-panel/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "@components/post-item",
+  "name": "@ui/search",
   "version": "0.0.0",
   "license": "BSD 3-Clause",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "scripts": {
     "lint": "eslint"
   },
   "dependencies": {
     "react": "^17.0.1",
-    "@components/post-list-item": "0.0.0"
+    "@ui/input": "0.0.0",
+    "@style/css": "0.0.0"
   }
 }
