+++ practice/web-diary/src/ui/search-panel/src/index.tsx
